src/PricingRegister/requirements.txt
google-cloud-bigquery==3.25.0
google-auth==2.32.0
urllib3==1.26.19
slack_sdk==3.31.0
